+++ .env
@@ -1,3 +1,3 @@
 DB_NAME='users_db'
-DB_PASSWORD='Bubbles11!'
+DB_PASSWORD='postgres'
 DB_USER='postgres'
