+++ controllers/api/index.js
@@ -1,13 +1,11 @@
 const router = require('express').Router();
 // Import the routes. This is how we make our routes modular.
 const userRoutes = require('./userRoutes');
-const projectRoutes = require('./projectRoutes');
 const eventRoutes   = require('./events');
 
 
 // When a request is made to the /users or /projects path, it will be directed to the index.js in the /users or /projects folder.
 router.use('/users', userRoutes);
 router.use('/events', eventRoutes);
-// router.use('/projects', projectRoutes);
 
 module.exports = router;
