+++ controllers/index.js
@@ -3,11 +3,11 @@ const router = require('express').Router();
 
 // Import the index.js from 'api' folder
 const apiRoutes = require('./api');
-const homeRoutes = require('./api/homeRoutes');
-// const userRoutes = require('./api/userRoutes');
+const homeRoutes = require('./homeRoutes');
+const userRoutes = require('./api/userRoutes');
 
 router.use('/api', apiRoutes);
 router.use('/', homeRoutes);
-// router.use('/users', userRoutes);
+router.use('/users', userRoutes);
 
 module.exports = router;
